+++ clients/marker.py
@@ -390,10 +390,18 @@ class ProxyModel(QSortFilterProxyModel):
 # TODO: should be a QMainWindow but at any rate not a Dialog
 # TODO: should this be parented by the QApplication?
 class MarkerClient(QWidget):
-    my_shutdown_signal = pyqtSignal(int)
+    my_shutdown_signal = pyqtSignal(int, list)
 
     def __init__(
-        self, userName, password, server, message_port, web_port, pageGroup, version
+        self,
+        userName,
+        password,
+        server,
+        message_port,
+        web_port,
+        pageGroup,
+        version,
+        lastTime,
     ):
         # Init the client with username, password, server and port data,
         # and which group/version is being marked.
@@ -446,7 +454,11 @@ class MarkerClient(QWidget):
         # create a settings variable for saving annotator window settings
         # initially all settings are "none"
         self.annotatorSettings = defaultdict(lambda: None)
-        # self.loadAnnotatorSettings()
+        # if lasttime["POWERUSER"] is true, the disable warnings in annotator
+        if "POWERUSER" in lastTime:
+            if lastTime["POWERUSER"]:
+                self.annotatorSettings["markWarnings"] = False
+                self.annotatorSettings["commentWarnings"] = False
 
         # Connect gui buttons to appropriate functions
         self.ui.closeButton.clicked.connect(self.shutDown)
@@ -470,12 +482,22 @@ class MarkerClient(QWidget):
         # continue with the other buttons
         self.ui.markStyleGroup.setId(self.ui.markUpRB, 2)
         self.ui.markStyleGroup.setId(self.ui.markDownRB, 3)
+        if lastTime["upDown"] == "up":
+            self.ui.markUpRB.animateClick()
+        elif lastTime["upDown"] == "down":
+            self.ui.markDownRB.animateClick()
+
         # Give IDs to the radio buttons which select which mouse-hand is used
         # 0 = Right-handed user will typically have right-hand on mouse and
         # left hand on the keyboard. The annotator layout will follow this.
         # 1 = Left-handed user - reverse layout
         self.ui.mouseHandGroup.setId(self.ui.rightMouseRB, 0)
         self.ui.mouseHandGroup.setId(self.ui.leftMouseRB, 1)
+        if lastTime["mouse"] == "right":
+            self.ui.rightMouseRB.animateClick()
+        elif lastTime["mouse"] == "left":
+            self.ui.leftMouseRB.animateClick()
+
         # Start using connection to serverself.
         try:
             self.token = requestToken(self.userName, self.password)
@@ -776,16 +798,16 @@ class MarkerClient(QWidget):
         questions are taking a long time or are quick).
         """
         # Set marking style total/up/down - will pass to annotator
-        self.markStyle = self.ui.markStyleGroup.checkedId()
+        markStyle = self.ui.markStyleGroup.checkedId()
         # Set mousehand left/right - will pass to annotator
-        self.mouseHand = self.ui.mouseHandGroup.checkedId()
+        mouseHand = self.ui.mouseHandGroup.checkedId()
         # Set plom-dictionary to none
         pdict = None
-        # check if given a plom-file and set markstyle + pdict accordingly
+        # check if given a plom-file and override markstyle + pdict accordingly
         if pname is not None:
             with open(pname, "r") as fh:
                 pdict = json.load(fh)
-            self.markStyle = pdict["markStyle"]
+            markStyle = pdict["markStyle"]
             # TODO: there should be a filename sanity check here to
             # make sure plom file matches current image-file
 
@@ -802,8 +824,8 @@ class MarkerClient(QWidget):
         annotator = Annotator(
             fname,
             self.maxScore,
-            self.markStyle,
-            self.mouseHand,
+            markStyle,
+            mouseHand,
             parent=self,
             plomDict=pdict,
         )
@@ -965,7 +987,7 @@ class MarkerClient(QWidget):
             self.updateImage(idx[0].row())
 
     def shutDownError(self):
-        self.my_shutdown_signal.emit(2)
+        self.my_shutdown_signal.emit(2, [])
         self.close()
 
     def shutDown(self):
@@ -1001,10 +1023,12 @@ class MarkerClient(QWidget):
         # authentication token.
         msg, = messenger.SRMsg(["UCL", self.userName, self.token])
         assert msg == "ACK"
-        # Now save annotatorSettings to file
-        # self.saveAnnotatorSettings()
+        # set marking style, mousehand for return to client/parent
+        markStyle = self.ui.markStyleGroup.checkedId()  # TODO
+        mouseHand = self.ui.mouseHandGroup.checkedId()
+
         # finally send shutdown signal to client window and close.
-        self.my_shutdown_signal.emit(2)
+        self.my_shutdown_signal.emit(2, [markStyle, mouseHand])
         self.close()
 
     def DNF(self):