+++ 00_new_project.py
@@ -53,8 +53,9 @@ files += [
     "build/cleanAll.py",
     "build/editMeToBuildASpec.py",
     "build/02_build_tests_from_spec.py",
+    "build/02a_build_with_names.py",
+    "build/build_utils.py",
     "build/mergeAndCodePages.py",
-    "build/buildTestPDFs.py",
     "build/ui_builder.py",
     "build/version_selector_qlist.py",
 ]