+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Warnings are given if some non-Latin-1 names are in the classlist (as those may cause problems in generated PDFs).
 
 ### Fixed
+* Annotator mark up/down and handedness preferences are saved between sessions.
 
 
 ## [0.2.1] - 2019-11-11