+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * Can now build papers with student Names/IDs pre-written on page 1.
+* Client now has a "view" button to quickly view other questions.
 
 ### Changed
 * Warning given for non-Latin names classlist (may cause problems in PDFs).